Refactor: Extract out `DecklistSelector`
part of #197
