+++ src/client/components/DeckListSelector/index.ts
@@ -1,0 +1,1 @@
+export * from './DeckListSelector';
